Replace legacy Lex/Yacc parser with shared cparser frontend
