+++ GPC/meson.build
@@ -6,7 +6,6 @@ libunwind_platform = compiler.find_library('unwind-x86_64', required: false)
 inc_dirs = include_directories(
   '.',
   'Parser',
-  'Parser/LexAndYacc',
   'Parser/ParseTree',
   'Parser/List',
   'Parser/SemanticCheck/HashTable',
@@ -41,6 +40,7 @@ gpc_cparser_sources = [
   'Parser/List/List.c',
   'Parser/ParseTree/tree.c',
   'Parser/parser_error.c',
+  'Parser/pascal_frontend.c',
   'Parser/SemanticCheck/SemCheck.c',
   'Parser/SemanticCheck/HashTable/HashTable.c',
   'Parser/SemanticCheck/SymTab/SymTab.c',
